adding more structure
