+++ api/v1/file/index.php
@@ -1,0 +1,5 @@
+<?php 
+
+import('api.v1.file.FileHandler');
+
+return  FileHandler::init();
